+++ src/ios/chrome/browser/crash_report/crash_helper.mm
@@ -7,6 +7,7 @@
 #import <UIKit/UIKit.h>
 #include <stddef.h>
 #include <stdint.h>
+#include <sys/stat.h>
 #include <sys/sysctl.h>
 
 #include "base/auto_reset.h"
@@ -16,18 +17,21 @@
 #include "base/files/file_enumerator.h"
 #include "base/files/file_path.h"
 #include "base/files/file_util.h"
+#include "base/ios/ios_util.h"
 #include "base/location.h"
 #include "base/logging.h"
 #include "base/metrics/histogram_functions.h"
+#include "base/metrics/histogram_macros.h"
 #include "base/path_service.h"
 #include "base/strings/sys_string_conversions.h"
-#include "base/task/post_task.h"
 #include "base/task/thread_pool.h"
+#include "base/time/time.h"
 #include "components/crash/core/app/crashpad.h"
 #include "components/crash/core/common/crash_key.h"
 #include "components/crash/core/common/reporter_running_ios.h"
 #include "ios/chrome/browser/chrome_paths.h"
 #import "ios/chrome/browser/crash_report/crash_report_user_application_state.h"
+#include "ios/chrome/browser/crash_report/crash_upload_list.h"
 #include "ios/chrome/browser/crash_report/features.h"
 #import "ios/chrome/browser/crash_report/main_thread_freeze_detector.h"
 #include "ios/chrome/common/app_group/app_group_constants.h"
@@ -46,13 +50,25 @@ namespace {
 const char kUptimeAtRestoreInMs[] = "uptime_at_restore_in_ms";
 const char kUploadedInRecoveryMode[] = "uploaded_in_recovery_mode";
 
-void DeleteAllReportsInDirectory(base::FilePath directory) {
+// Delete breakpad reports after 60 days.
+void DeleteOldReportsInDirectory(base::FilePath directory) {
   base::FileEnumerator enumerator(directory, false,
                                   base::FileEnumerator::FILES);
   base::FilePath cur_file;
   while (!(cur_file = enumerator.Next()).value().empty()) {
-    if (cur_file.BaseName().value() != kReporterLogFilename)
-      base::DeleteFile(cur_file);
+    if (cur_file.BaseName().value() != kReporterLogFilename) {
+      time_t now = time(nullptr);
+      struct stat st;
+      if (lstat(cur_file.value().c_str(), &st) != 0) {
+        continue;
+      }
+
+      // 60 days.
+      constexpr time_t max_breakpad_report_age_sec = 60 * 60 * 24 * 60;
+      if (st.st_mtime <= now - max_breakpad_report_age_sec) {
+        base::DeleteFile(cur_file);
+      }
+    }
   }
 }
 
@@ -114,7 +130,7 @@ int64_t GetUptimeMilliseconds() {
   kinfo_proc kern_proc_info;
   int mib[] = {CTL_KERN, KERN_PROC, KERN_PROC_PID, getpid()};
   size_t len = sizeof(kern_proc_info);
-  if (sysctl(mib, base::size(mib), &kern_proc_info, &len, nullptr, 0) != 0)
+  if (sysctl(mib, std::size(mib), &kern_proc_info, &len, nullptr, 0) != 0)
     return 0;
   time_t process_uptime_seconds =
       tv.tv_sec - kern_proc_info.kp_proc.p_starttime.tv_sec;
@@ -138,11 +154,14 @@ void Start() {
   // can be displayed to the user on the about:crashes page.  Use the app group
   // so crashes can be shared by plugins.
   logging::SetLogMessageHandler(&FatalMessageHandler);
-  if (common::CanCrashpadStart()) {
+  if (common::CanUseCrashpad()) {
     base::PathService::Override(ios::DIR_CRASH_DUMPS,
                                 common::CrashpadDumpLocation());
-    common::StartCrashpad();
-    crash_reporter::SetCrashpadRunning(true);
+    bool initialized = common::StartCrashpad();
+    if (initialized) {
+      crash_reporter::SetCrashpadRunning(true);
+    }
+    UMA_HISTOGRAM_BOOLEAN("Stability.IOS.Crashpad.Initialized", initialized);
   } else {
     NSArray* cachesDirectories = NSSearchPathForDirectoriesInDomains(
         NSCachesDirectory, NSUserDomainMask, YES);
@@ -163,16 +182,35 @@ void Start() {
       key.Set(channel_name);
     }
   }
-  [[MainThreadFreezeDetector sharedInstance] start];
+
+  crash_reporter::InitializeCrashKeys();
+
+  // Don't start MTFD when prewarmed, the check thread will just get confused.
+  if (!base::ios::IsApplicationPreWarmed()) {
+    [[MainThreadFreezeDetector sharedInstance] start];
+  }
 }
 
 void SetEnabled(bool enabled) {
+  // Caches the uploading flag in NSUserDefaults, so that we can access the
+  // value immediately on startup, such as in safe mode or extensions.
+  crash_helper::common::SetUserEnabledUploading(enabled);
+
   // It is necessary to always call |MainThreadFreezeDetector setEnabled| as
   // the function will update its preference based on finch.
   [[MainThreadFreezeDetector sharedInstance] setEnabled:enabled];
 
-  // Crashpad is always running, don't shut it off.
-  if (crash_reporter::IsCrashpadRunning()) {
+  // Crashpad is always running, don't shut it off. Using CanUseCrashpad()
+  // here, because if Crashpad fails to init, do not unintentionally enable
+  // breakpad.
+  if (common::CanUseCrashpad()) {
+    // Posts SetUploadConsent on blocking pool thread because it needs access to
+    // IO and cannot work from UI thread.
+    base::ThreadPool::PostTask(
+        FROM_HERE, {base::MayBlock(), base::TaskPriority::BEST_EFFORT},
+        base::BindOnce(^{
+          crash_reporter::SetUploadConsent(enabled);
+        }));
     return;
   }
 
@@ -186,66 +224,38 @@ void SetEnabled(bool enabled) {
   }
 }
 
-// Breakpad only.
-void SetBreakpadUploadingEnabled(bool enabled) {
-  if (!crash_reporter::IsBreakpadRunning())
-    return;
-  if (enabled) {
+void UploadCrashReports() {
+  if (crash_reporter::IsCrashpadRunning()) {
+    static dispatch_once_t once_token;
+    dispatch_once(&once_token, ^{
+      base::ThreadPool::PostTask(
+          FROM_HERE, {base::MayBlock(), base::TaskPriority::BEST_EFFORT},
+          base::BindOnce(&ProcessIntermediateDumps));
+      return;
+    });
+  }
+
+  if (crash_reporter::IsBreakpadRunning()) {
     static dispatch_once_t once_token;
     dispatch_once(&once_token, ^{
       [[BreakpadController sharedInstance]
           setUploadCallback:UploadResultHandler];
+
+      // Clean old breakpad files here. Breakpad-only as Crashpad has it's own
+      // database cleaner.
+      base::FilePath crash_directory;
+      base::PathService::Get(ios::DIR_CRASH_DUMPS, &crash_directory);
+      base::ThreadPool::PostTask(
+          FROM_HERE, {base::MayBlock(), base::TaskPriority::BEST_EFFORT},
+          base::BindOnce(&DeleteOldReportsInDirectory, crash_directory));
     });
-  }
-  [[BreakpadController sharedInstance] setUploadingEnabled:enabled];
-}
-
-// Caches the uploading flag in NSUserDefaults, so that we can access the value
-// immediately on startup, such as in safe mode or extensions.
-void SetUserEnabledUploading(bool uploading_enabled) {
-  [app_group::GetGroupUserDefaults()
-      setBool:uploading_enabled ? YES : NO
-       forKey:base::SysUTF8ToNSString(
-                  common::kCrashReportsUploadingEnabledKey)];
-}
-
-void SetUploadingEnabled(bool enabled) {
-  if (enabled && [UIApplication sharedApplication].applicationState ==
-                     UIApplicationStateInactive) {
-    return;
-  }
-
-  if (crash_reporter::IsCrashpadRunning()) {
-    crash_reporter::SetUploadConsent(enabled);
-    [[MainThreadFreezeDetector sharedInstance] prepareCrashReportsForUpload:^(){
-    }];
-    return;
-  }
-
-  if ([MainThreadFreezeDetector sharedInstance].canUploadBreakpadCrashReports) {
-    SetBreakpadUploadingEnabled(enabled);
-  } else {
-    [[MainThreadFreezeDetector sharedInstance]
-        prepareCrashReportsForUpload:^() {
-          SetBreakpadUploadingEnabled(enabled);
-        }];
+    [[BreakpadController sharedInstance] setUploadingEnabled:YES];
   }
 }
 
-void CleanupCrashReports(BOOL after_upgrade) {
-  if (crash_reporter::IsCrashpadRunning()) {
-    base::ThreadPool::PostTask(
-        FROM_HERE, {base::MayBlock(), base::TaskPriority::BEST_EFFORT},
-        base::BindOnce(&ProcessIntermediateDumps));
-    return;
-  }
-
-  if (after_upgrade) {
-    base::FilePath crash_directory;
-    base::PathService::Get(ios::DIR_CRASH_DUMPS, &crash_directory);
-    base::ThreadPool::PostTask(
-        FROM_HERE, {base::MayBlock(), base::TaskPriority::BEST_EFFORT},
-        base::BindOnce(&DeleteAllReportsInDirectory, crash_directory));
+void PauseBreakpadUploads() {
+  if (crash_reporter::IsBreakpadRunning()) {
+    [[BreakpadController sharedInstance] setUploadingEnabled:NO];
   }
 }
 
@@ -279,14 +289,6 @@ int GetPendingCrashReportCount() {
   }];
   dispatch_semaphore_wait(semaphore, DISPATCH_TIME_FOREVER);
   return outerCrashReportCount;
-}
-
-void GetCrashReportCount(void (^callback)(int)) {
-  if (crash_reporter::IsCrashpadRunning()) {
-    callback(GetPendingCrashReportCount());
-  }
-
-  [[BreakpadController sharedInstance] getCrashReportCount:callback];
 }
 
 bool HasReportToUpload() {
@@ -354,16 +356,18 @@ void StartUploadingReportsInRecoveryMode() {
 }
 
 void RestoreDefaultConfiguration() {
-  if (crash_reporter::IsCrashpadRunning()) {
+  if (!crash_reporter::IsBreakpadRunning()) {
     return;
   }
 
-  if (!crash_reporter::IsBreakpadRunning())
-    return;
   [[BreakpadController sharedInstance] stop];
   [[BreakpadController sharedInstance] resetConfiguration];
   [[BreakpadController sharedInstance] start:NO];
   [[BreakpadController sharedInstance] setUploadingEnabled:NO];
+}
+
+void ClearReportsBetween(base::Time delete_begin, base::Time delete_end) {
+  ios::CreateCrashUploadList()->Clear(delete_begin, delete_end);
 }
 
 }  // namespace crash_helper
